Default configuration of arguments for eemanalyzeR

The default argument values used in the run_eems() function. Argument values are stored in a package environment (.pkgenv) which will be accessed in the case where a specific variable is not specified by the user.

Usage

default_config

Format

A list of length 27:

abs_pattern: Used by abs_dir_read(). A character string containing a base::regular expression() to match files in input_dir. Only files matching the pattern will be loaded.

abs_skip: Used by abs_dir_read(). A character string containing a base::regular expression() to match files in input_dir that should be ignored.

abs_file_ext: Used by abs_dir_read(). The file extension of the absorbance files.

abs_recurse_read: Used by eem_dir_read(). Logical. Should the function recursively search directories?

eem_pattern: Used by eem_dir_read(). A character string containing a base::regular expression() to match files in input_dir. Only files matching the pattern will be loaded.

eem_skip: Used by eem_dir_read(). A character string containing a base::regular expression() to match files in input_dir that should be ignored.

eem_file_ext: Used by eem_dir_read(). The file extension of the EEMs files.

eem_recurse_read: Used by eem_dir_read(). Logical. Should the function recursively search directories?

eem_import_func: Used by abs_dir_read(). Character or a user-defined function to import an EEM. For more details, see vignette("custom-indices").

meta_sheet: Used by meta_read(). Name of the sheet containing metadata (only required if the metadata is not on the first sheet of an .xlsx file).

meta_validate: Used by meta_read(). Logical. If TRUE, checks the metadata for structural issues that could cause problems during processing. Recommended to keep TRUE.

iblank_pattern: Used by add_metadata(). A character vector of length 1 with a regular expression that matches sample names of instrument blanks.

sblank_pattern: Used by add_metadata(). A character vector of length 1 with a regular expression that matches sample names of sample blanks.

check_pattern: Used by add_metadata(). A character vector of length 1 with a regular expression that matches sample names of check standards.

blank_validate: Used by add_blanks(). Logical vector length one indicating whether blanks should be validated.

ex_clip: Used by process_eem(). Numeric vector of length two specifying the minimum and maximum excitation wavelengths to keep.

em_clip: Used by process_eem(). Numeric vector of length two specifying the minimum and maximum emission wavelengths to keep.

type: Used by process_eem(). Logical vector of length four indicating which scattering lines to remove. The order is "raman1", "raman2", "rayleigh1", "rayleigh2".

width: Used by process_eem(). Numeric vector of length four specifying the width of scattering lines to remove (nm). Same order as type.

interpolate: Used by process_eem(). Logical vector of length four indicating which scattering lines to interpolate. Same order as type.

method: Used by process_eem(). Numeric (0–4) specifying the interpolation method to use. Default is 1. See staRdom::eem_interp() for details.

cores: Used by process_eem(). Integer specifying the number of cores for parallel computation during interpolation.

cuvle: Used by process_eem(). Cuvette (path) length in cm.

index_method: Used by get_indices(). Either "eemanalyzeR", "eemR", "usgs", or a custom function.

tolerance: Used by get_indices(). Maximum percent deviation that the check standard can vary from the long-term values without being flagged.

return: Used by get_indices().Output format: "long" or "wide".

qaqc_dir: Used by get_indices(). File path to the QAQC files generated with create_mdl() and create_std(). Default is a user-specific data directory rappdirs::user_data_dir().

filename: Used by export_data(). A character string, used for file names.

output_dir: Used by export_data(). Path to save the data. Defaults to a temporary directory if not specified.

readme: Starts as NULL, used to store notes and warnings about processing.

Details

These values can be edited by:

the user permanently using edit_user_config()

the user for the current R session using modify_config()

or within the run_eems() function itself by providing argument values