Standard ranges for absorbance and EEMs indices

A data.frame containing the minimum and maximum range expected for index values generated using get_indices(). Used to flag indices that warrant further examination to ensure accuracy.

Usage

indice_ranges

Format

A data.frame with 62 rows and 5 columns:

index_method: The index method used to calculate the index

index: Name of the index metric calculated

low_val: Minimum value expected based on literature sources

high_val: Maximum value expected based on literature sources

source: Sources used to determine minimum and maximum values

Source

Fischer, S. J., Fegel, T. S., Wilkerson, P. J., Rivera, L., Rhoades, C. C., & Rosario-Ortiz, F. L. (2023). Fluorescence and Absorbance Indices for Dissolved Organic Matter from Wildfire Ash and Burned Watersheds. ACS ES&T Water, 3(8), 2199-2209. https://doi.org/10.1021/acsestwater.3c00017

Galgani, L., Engel, A., Rossi, C., Donati, A., & Loiselle, S. A. (2018). Polystyrene microplastics increase microbial release of marine Chromophoric Dissolved Organic Matter in microcosm experiments. Scientific Reports, 8(1), 14635. https://doi.org/10.1038/s41598-018-32805-4

Hansen, A. M., Kraus, T. E. C., Pellerin, B. A., Fleck, J. A., Downing, B. D., & Bergamaschi, B. A. (2016). Optical properties of dissolved organic matter (DOM): Effects of biological and photolytic degradation. Limnology and Oceanography, 61(3), 1015-1032. https://doi.org/10.1002/lno.10270

Hansen, A. M., Fleck, J., Kraus, T. E. C., Downing, B. D., von Dessonneck, T., & Bergamaschi, B. (2018). Procedures for using the Horiba Scientific Aqualog® fluorometer to measure absorbance and fluorescence from dissolved organic matter (USGS Numbered Series No. 2018-1096). Procedures for using the Horiba Scientific Aqualog® fluorometer to measure absorbance and fluorescence from dissolved organic matter (Vol. 2018-1096). Reston, VA: U.S. Geological Survey. https://doi.org/10.3133/ofr20181096

Helms, J. R., Stubbins, A., Ritchie, J. D., Minor, E. C., Kieber, D. J., & Mopper, K. (2008). Absorption spectral slopes and slope ratios as indicators of molecular weight, source, and photobleaching of chromophoric dissolved organic matter. Limnology and Oceanography, 53(3), 955-969. https://doi.org/10.4319/lo.2008.53.3.0955

Korak, J. A., & McKay, G. (2024). Meta-Analysis of Optical Surrogates for the Characterization of Dissolved Organic Matter. Environmental Science & Technology, 58(17), 7380-7392. https://doi.org/10.1021/acs.est.3c10627

Li, Y., Zhang, Y., Li, Z., Wan, J., Dang, C., & Fu, J. (2022). Characterization of colored dissolved organic matter in the northeastern South China Sea using EEMs-PARAFAC and absorption spectroscopy. Journal of Sea Research, 180, 102159. https://doi.org/10.1016/j.seares.2021.102159

Meingast, K. M., Kane, E. S., Marcarelli, A. M., Wagenbrenner, J. W., & Beltrone, V. G. (2023). Seasonal trends of DOM character in soils and stream change with snowmelt timing. Water Resources Research. 59(3): e2022WR032014. https://doi.org/10.1029/2022WR032014

Peuravuori, J., & Pihlaja, K. (1997). Molecular size distribution and spectroscopic properties of aquatic humic substances. Analytica Chimica Acta, 337(2), 133-149. https://doi.org/10.1016/S0003-2670(96)00412-6

Weishaar, J. L., Aiken, G. R., Bergamaschi, B. A., Fram, M. S., Fujii, R., & Mopper, K. (2003). Evaluation of Specific Ultraviolet Absorbance as an Indicator of the Chemical Composition and Reactivity of Dissolved Organic Carbon. Environmental Science & Technology, 37(20), 4702-4708. https://doi.org/10.1021/es030360x

Zalba, P., Amiotti ,Nilda M., Galantini ,Juan A., & and Pistola, S. (2016). Soil Humic and Fulvic Acids from Different Land-Use Systems Evaluated By E4/E6 Ratios. Communications in Soil Science and Plant Analysis, 47(13-14), 1675-1679. https://doi.org/10.1080/00103624.2016.1206558

Zsolnay, A., Baigar, E., Jimenez, M., Steinweg, B., & Saccomandi, F. (1999). Differentiating with fluorescence spectroscopy the sources of dissolved organic matter in soils subjected to drying. Chemosphere, 38(1), 45-50. https://doi.org/10.1016/S0045-6535(98)00166-0